final du tp calculette avec flottant et division

+++ ast_construction.c
@@ -24,7 +24,7 @@ Ast creer_operation(TypeOperateur opr , Ast opde_gauche , Ast opde_droit)
       return expr ;
 }
 
-Ast creer_valeur(int val) { 
+Ast creer_valeur(float val) { 
       Ast expr ;
       expr = (Ast) malloc (sizeof(NoeudAst));
       expr->nature = VALEUR;
