+++ ast_parcours.h
@@ -4,7 +4,7 @@
 void afficher_ast(Ast expr) ;
 // affiche l'arbre abstrait de l'expression arithmetique expr
 
-int evaluation(Ast expr) ;
+float evaluation(Ast expr) ;
 // calcule la valeur de l'expression arithmetique expr
 // FONCTION A COMPLETER !
